fanya-ai-backend/requirements.txt
Flask==2.0.1
Flask-CORS==3.0.10
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.46
Flask-JWT-Extended==4.4.4
PyMySQL==1.0.2
passlib==1.7.4
python-dotenv==0.21.0
Werkzeug==2.0.3

# TTS (必须)
ChatTTS>=0.1.0
torch>=2.1.0
torchaudio>=2.1.0

# AI 文本生成 (可选，根据需要安装)
# 智谱 GLM API (推荐)
zhipuai>=0.0.6

# OpenAI ChatGPT API
openai>=0.27.0

# 阿里通义千问 API
dashscope>=1.0.0

# 讯飞星火相关依赖
requests>=2.28.0
websocket-client>=1.0.0


# Gradio 客户端
gradio-client==0.10.1

# PPT转PDF
comtypes>=1.2.0
reportlab>=4.0.0

# RAG / 文档解析增强
jieba
rank-bm25
# 固定版本避免 pip 在 docling 多版本上长时间回溯
docling==2.79.0
python-pptx
PyPDF2
edge-tts